refactor(kernel,backends): Phase D start — lift LlmStreamingBackend trait + add backends crate skeleton
Two related Phase D prep changes packaged together:

1) Lift `LlmStreamingBackend` trait declaration from
   `openai_streaming.rs` to `core/traits/llm_streaming.rs`. Phase D
   will move openai_streaming.rs into `backends/src/transports/ai/openai/`,
   which means ObjectStore::as_llm_streaming returning
   &dyn LlmStreamingBackend in the kernel trait would otherwise
   reference a type living in the parallel backends/ crate — a
   kernel→backends dependency that, combined with the existing
   backends→kernel dep on ObjectStore, forms a cycle. The trait stays
   kernel-side; the impl moves with the connector. Mirrors the plan's
   PeerBlobClient pattern for the peer_client move (Phase F).
   `crate::openai_streaming::LlmStreamingBackend` keeps resolving via a
   pub use re-export so callers do not churn.

2) Add empty `rust/backends/` crate. Workspace member +
   `backends = { path = "rust/backends" }` workspace dep, plus a
   workspace dep entry for `kernel = { path = "rust/kernel",
   default-features = false }` so backends can pull the kernel rlib
   without the cdylib's `extension-module` feature leaking. Cargo.toml
   carries the same connector / nostr / python feature gates kernel's
   own does today; `src/lib.rs` is currently a single
   `#[cfg(feature = "python")] pub mod python;` placeholder. The
   subsequent commits move the actual driver impls in.

scripts/codegen_kernel_abi.py picked up the new trait — kernel_protocols.py
gains a `LlmStreamingBackend(Protocol)` class.

cargo check --workspace clean. No behavior change.
